Re-route comment, create another form for Admin\CommentType
