+++ src/Controller/Admin/CommentController.php
@@ -5,16 +5,55 @@ namespace App\Controller\Admin;
 use App\Repository\CommentRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use App\Entity\Comment;
+use App\Form\Admin\CommentType;
 
 #[Route('/admin/comment')]
 class CommentController extends AbstractController
 {
-    #[Route('/', name: 'app_admin_comment')]
+    #[Route('/', name: 'app_admin_comment_index', methods: ['GET'])]
     public function index(CommentRepository $commentRepository): Response
     {
         return $this->render('admin/comment/index.html.twig', [
             'comments' => $commentRepository->findAll(),
         ]);
+    }
+
+    #[Route('/{id}/edit', name: 'app_comment_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Comment $comment, CommentRepository $commentRepository): Response
+    {
+        $form = $this->createForm(CommentType::class, $comment);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $commentRepository->save($comment, true);
+
+            return $this->redirectToRoute('app_admin_comment_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('admin/comment/edit.html.twig', [
+            'comment' => $comment,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'app_comment_show', methods: ['GET'])]
+    public function show(Comment $comment): Response
+    {
+        return $this->render('admin/comment/show.html.twig', [
+            'comment' => $comment,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'app_comment_delete', methods: ['POST'])]
+    public function delete(Request $request, Comment $comment, CommentRepository $commentRepository): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$comment->getId(), $request->request->get('_token'))) {
+            $commentRepository->remove($comment, true);
+        }
+
+        return $this->redirectToRoute('app_admin_comment_index', [], Response::HTTP_SEE_OTHER);
     }
 }
