+++ src/Controller/PostController.php
@@ -5,10 +5,8 @@ namespace App\Controller;
 use App\Entity\Comment;
 use App\Form\CommentType;
 use App\Entity\Post;
-use App\Form\PostType;
 use App\Repository\CategoryRepository;
 use App\Repository\CommentRepository;
-use App\Repository\PostRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
