+++ src/Entity/User.php
@@ -43,6 +43,11 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
         $this->comments = new ArrayCollection();
     }
 
+    public function __toString(): string
+    {
+        return $this->username;
+    }
+
     public function getId(): ?int
     {
         return $this->id;
