+++ src/Form/Admin/CommentType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Admin;
 
 use App\Entity\Comment;
 use Symfony\Component\Form\AbstractType;
@@ -13,11 +13,11 @@ class CommentType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('content', TextareaType::class, array('label' => 'Comments'));
-            // ->add('owner');
-            // ->add('valid')
-            // ->add('createdAt')
-            // ->add('post')
+            ->add('content', TextareaType::class, array('label' => 'Comments'))
+            ->add('owner')
+            ->add('valid')
+            ->add('createdAt')
+            ->add('post');
 
     }
