+++ templates/admin/admin/index.html.twig
@@ -11,7 +11,7 @@
                     <li><a href="{{ path('app_admin_user_index') }}">Users</a></li>
                     <li><a href="{{ path('app_admin_category') }}">Categories</a></li>
                     <li><a href="{{ path('app_admin_post_index') }}">Posts</a></li>
-                    <li><a href="{{ path('app_admin_comment') }}">Comments</a></li>
+                    <li><a href="{{ path('app_admin_comment_index') }}">Comments</a></li>
                 </ul>
             </aside>
         </section>
