+++ templates/admin/comment/_form.html.twig
@@ -1,0 +1,4 @@
+{{ form_start(form) }}
+    {{ form_widget(form) }}
+    <button class="btn btn-dark mt-5">{{ button_label|default('Post Comment') }}</button>
+{{ form_end(form) }}
