+++ templates/admin/comment/edit.html.twig
@@ -1,0 +1,11 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Edit Comment{% endblock %}
+
+{% block body %}
+    <h1>Edit Comment</h1>
+
+    {{ include('admin/comment/_form.html.twig', {'button_label': 'Update'}) }}
+
+    {{ include('admin/comment/_delete_form.html.twig') }}
+{% endblock %}
