+++ templates/admin/comment/show.html.twig
@@ -26,9 +26,7 @@
         </tbody>
     </table>
 
-    <a href="{{ path('app_comment_index') }}">back to list</a>
+    <a href="{{ path('app_comment_edit', {'id': comment.id}) }}"><button type="button" class="btn btn-dark">Edit</button></a>
 
-    <a href="{{ path('app_comment_edit', {'id': comment.id}) }}">edit</a>
-
-    {{ include('comment/_delete_form.html.twig') }}
+    {{ include('admin/comment/_delete_form.html.twig') }}
 {% endblock %}
