--- templates/comment/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Edit Comment{% endblock %}
-
-{% block body %}
-    <h1>Edit Comment</h1>
-
-    {{ include('comment/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('app_comment_index') }}">back to list</a>
-
-    {{ include('comment/_delete_form.html.twig') }}
-{% endblock %}
